Added a bunch of stuff for sparse vector and matrix
+ Added SparseDoubleVector, SparseGrowableDoubleVector,
SparseGrowableVector
+ Added SparseArrayMatrix, SparseGrowableMatrix and SparseDoubleMatrix.
+ Added tests for vectors.
+ At this point we still run out of memory trying to initialize the
matrix from the PageLink structure; we can’t have these both in memory
at the same time. So we need to restructure how we read from the input
file so that it is sorted by fromPage so we can read all the associated
toPage and set them into the matrix.  We can do this if we have all the
toPage values for the fromPage; if we do this then we only need to read
the link from one page at a time; we don’t need them all in memory at
once.